+++ sources/afx/video/afx_video_mjpeg+/XJpegHttpStream.cpp
@@ -330,7 +330,7 @@ void XJpegHttpStream::RunVideo( )
                     // because found a camera, which keeps rejecting libcurl's authorization
 
                     string  loginWithPassword = mData->UserName + ":" + mData->Password;
-                    int     authLen           = Private::Base64EncodeLength( loginWithPassword.size( ) + 21 );
+                    size_t  authLen           = Private::Base64EncodeLength( loginWithPassword.size( ) + 21 );
                     char*   authorization     = new char[authLen];
 
                     strcpy( authorization, "Authorization: Basic " );
@@ -602,7 +602,7 @@ namespace Private
         if ( realSize <= MaxBufferSize - data->ReadSoFar )
         {
             memcpy( &(data->CommunicationBuffer[data->ReadSoFar]), contents, realSize );
-            data->ReadSoFar += realSize;
+            data->ReadSoFar += static_cast<uint32_t>( realSize );
         }
         else
         {
